+++ src/extensions/mvs/load-helpers.ts
@@ -26,7 +26,7 @@ import { ElementOfSet, decodeColor, isDefined, stringHash } from './helpers/util
 import { MolstarLoadingContext } from './load';
 import { Subtree, getChildren } from './tree/generic/tree-schema';
 import { dfs, formatObject } from './tree/generic/tree-utils';
-import { MolstarKind, MolstarNode, MolstarSubtree, MolstarTree, MolstarNodeParams } from './tree/molstar/molstar-tree';
+import { MolstarKind, MolstarNode, MolstarNodeParams, MolstarSubtree, MolstarTree } from './tree/molstar/molstar-tree';
 import { DefaultColor } from './tree/mvs/mvs-defaults';
 
 
@@ -97,7 +97,7 @@ export function collectAnnotationReferences(tree: Subtree<MolstarTree>, context:
         if (spec) {
             const key = canonicalJsonString(spec as any);
             distinctSpecs[key] ??= { ...spec, id: stringHash(key) };
-            (context.annotationMap ??= new Map()).set(node, distinctSpecs[key].id);
+            context.annotationMap.set(node as MolstarNode<AnnotationFromUriKind | AnnotationFromSourceKind>, distinctSpecs[key].id);
         }
     });
     return Object.values(distinctSpecs);
@@ -115,7 +115,7 @@ export function collectAnnotationTooltips(tree: MolstarSubtree<'structure'>, con
     const annotationTooltips: MVSAnnotationTooltipsProps['tooltips'] = [];
     dfs(tree, node => {
         if (node.kind === 'tooltip_from_uri' || node.kind === 'tooltip_from_source') {
-            const annotationId = context.annotationMap?.get(node);
+            const annotationId = context.annotationMap.get(node);
             if (annotationId) {
                 annotationTooltips.push({ annotationId, fieldName: node.params.field_name });
             };
@@ -259,7 +259,7 @@ export function prettyNameFromSelector(selector?: MolstarNodeParams<'component'>
 
 /** Create props for `StructureRepresentation3D` transformer from a label_from_* node. */
 export function labelFromXProps(node: MolstarNode<'label_from_uri' | 'label_from_source'>, context: MolstarLoadingContext): Partial<StateTransformer.Params<StructureRepresentation3D>> {
-    const annotationId = context.annotationMap?.get(node);
+    const annotationId = context.annotationMap.get(node);
     const fieldName = node.params.field_name;
     const nearestReprNode = context.nearestReprMap?.get(node);
     return {
@@ -270,7 +270,7 @@ export function labelFromXProps(node: MolstarNode<'label_from_uri' | 'label_from
 
 /** Create props for `AnnotationStructureComponent` transformer from a component_from_* node. */
 export function componentFromXProps(node: MolstarNode<'component_from_uri' | 'component_from_source'>, context: MolstarLoadingContext): Partial<MVSAnnotationStructureComponentProps> {
-    const annotationId = context.annotationMap?.get(node);
+    const annotationId = context.annotationMap.get(node);
     const { field_name, field_values } = node.params;
     return {
         annotationId,
@@ -302,12 +302,11 @@ export function representationProps(node: MolstarSubtree<'representation'>): Par
     }
 }
 
-/** Create value for `type.params.alpha` prop for `StructureRepresentation3D` transformer from a representation node based on 'transparency' nodes in its subtree. */
+/** Create value for `type.params.alpha` prop for `StructureRepresentation3D` transformer from a representation node based on 'opacity' nodes in its subtree. */
 export function alphaForNode(node: MolstarSubtree<'representation'>): number {
-    const children = getChildren(node).filter(c => c.kind === 'transparency');
+    const children = getChildren(node).filter(c => c.kind === 'opacity');
     if (children.length > 0) {
-        const transparency = children[children.length - 1].params.transparency;
-        return 1 - transparency;
+        return children[children.length - 1].params.opacity;
     } else {
         return 1;
     }
@@ -339,7 +338,7 @@ export function colorThemeForNode(node: MolstarSubtree<'color' | 'color_from_uri
     switch (node?.kind) {
         case 'color_from_uri':
         case 'color_from_source':
-            annotationId = context.annotationMap?.get(node);
+            annotationId = context.annotationMap.get(node);
             fieldName = node.params.field_name;
             break;
         case 'color':