+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- Fix outlines on transparent background using illumination mode (#1364)
+- Fix transparent depth texture artifacts using illumination mode
+ 
 ## [v4.9.1] - 2024-12-05
 
 - Fix iOS check when running on Node