+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add `ModelWithCoordinates` decorator transform.
 - Fix outlines on transparent background using illumination mode (#1364)
 - Fix transparent depth texture artifacts using illumination mode
+- Fix marking of consecutive gap elements (#876)
 - Allow React 19 in dependencies
 - Fix missing deflate header if `CompressionStream` is available