+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Added PluginCommands.Camera.FocusObject
 - Plugin state snapshot can have instructions to focus objects (PluginState.Snapshot.camera.focus)
 - MolViewSpec extension: Support for multi-state files (animations)
+- Fix units transform data not fully updated when structure child changes
 
 ## [v4.9.1] - 2024-12-05